Handle multiple TVs correctly
We have to lookup device from passed-in entity_id rather than from
bound "device" variable.

The registration of services is only really need to be done once but
it's still done multiple times if multiple bravia platforms are defined.
I haven't changed that as HA seems to handle it itself and doesn't
register duplicate services (it overrides the previous one).

Fixes #37